finished i think

+++ Tipper/app/src/main/java/com/uottawa/tipper/BillAmntFragment.java
@@ -114,7 +114,7 @@ public class BillAmntFragment extends Fragment {
                     String tipValue = String.valueOf(bill.getText()).trim();
 
                     if (tipValue.length() != 0 && !numberIn){
-                        double billamnt = Integer.parseInt(((EditText) rootView.findViewById(R.id.totalBillAmnt)).getText().toString());
+                        double billamnt = Double.parseDouble(((EditText) rootView.findViewById(R.id.totalBillAmnt)).getText().toString());
                         arrow.setTypeface(fontAwesome);
                         arrow.setTextColor(Color.parseColor("#32A0A0"));
                         arrow.startAnimation(AnimationUtils.loadAnimation(getActivity(),android.R.anim.slide_in_left));
